Log viewer implemented

+++ utils/logger/KLog.cs
@@ -1,3 +1,4 @@
+using System;
 using MongoDB.Bson;
 using MongoDB.Bson.Serialization.Attributes;
 
@@ -9,7 +10,7 @@ namespace dc.assignment.primenumbers.utils.logger
         [BsonRepresentation(BsonType.ObjectId)]
         public string? Id { get; set; }
         public string timestamp { get; set; } = null!;
-        public int nodeId { get; set; }
+        public Int64 nodeId { get; set; }
         public string nodeAddress { get; set; } = null!;
         public string message { get; set; } = null!;
     }
